requirements.txt
djangorestframework
psycopg2
python-decouple
nanoid